CCM-4894: rename file
