+++ src/__tests__/components/forms/CreateSmsTemplate/view-actions.test.ts
@@ -1,4 +1,4 @@
-import { calculateHowManySmsMessages } from '@forms/CreateSmsTemplate/form-actions';
+import { calculateHowManySmsMessages } from '@forms/CreateSmsTemplate/view-actions';
 
 describe('calculateHowManySmsMessages', () => {
   test.each([
